requirements.txt
fastapi
joblib
requests
pandas
numpy
scipy
implicit
scikit-learn
pyarrow
python-dotenv